leafdevice/requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

azure-iot-device==2.5.1

certifi==2020.12.5

chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

deprecation==2.1.0

idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

janus==0.4.0; python_version >= '3.5'

packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

paho-mqtt==1.5.1

pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

python-dotenv==0.17.0

requests-unixsocket==0.2.0

requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

urllib3==1.25.11; python_version != '3.4'

